libcoap_rs::crypto::psk

Trait ServerPskSniKeyProvider

pub trait ServerPskSniKeyProvider<'a>: Debug {
    // Required method
    fn key_for_sni(
        &self,
        sni: &CStr,
        session: &CoapServerSession<'_>,
    ) -> Option<PskKey<'a>>;
}

Trait for things that can provide PSK DTLS keys for a given Server Name Indication.

Required Methods

fn key_for_sni( &self, sni: &CStr, session: &CoapServerSession<'_>, ) -> Option<PskKey<'a>>

Provide a key for the server name indication given as sni, or None if the SNI is not valid and no key is available.

Note that libcoap will remember the returned key and re-use it for future handshakes with the same SNI (even if the peer is not the same), the return value should therefore not depend on the provided session.