libcoap_rs

Module message

Types related to message handling, parsing and creation.

The base unit that is transmitted between a CoAP client and a CoAP server is called a CoAP message (in libcoap also referred to as a “pdu”). This module contains both the basic representation for messages in libcoap-rs (CoapMessage) as well as builders that simplify the process of creating requests and responses and setting the appropriate options (CoapRequest and CoapResponse).

Re-exports

pub use request::CoapRequest;

pub use response::CoapResponse;

Modules

request

response

Structs

CoapMessage

Representation of a CoAP message.

Enums

CoapOption

Representation of a CoAP option including its value.

Traits

CoapMessageCommon

Interface for CoAP messages common between requests, responses and other messages.