libcoap_rs::message
Trait CoapMessageCommon

pub trait CoapMessageCommon {

    // Required methods
    fn as_message(&self) -> &CoapMessage;
    fn as_message_mut(&mut self) -> &mut CoapMessage;

    // Provided methods
    fn add_option(&mut self, option: CoapOption) { ... }
    fn clear_options(&mut self) { ... }
    fn options_iter(&self) -> Iter<'_, CoapOption> { ... }
    fn type_(&self) -> CoapMessageType { ... }
    fn set_type_(&mut self, type_: CoapMessageType) { ... }
    fn code(&self) -> CoapMessageCode { ... }
    fn set_code<C: Into<CoapMessageCode>>(&mut self, code: C) { ... }
    fn mid(&self) -> Option<CoapMessageId> { ... }
    fn set_mid(&mut self, mid: Option<CoapMessageId>) { ... }
    fn data(&self) -> Option<&[u8]> { ... }
    fn set_data<D: Into<Box<[u8]>>>(&mut self, data: Option<D>) { ... }
    fn token(&self) -> Option<&[u8]> { ... }
    fn set_token<D: Into<Box<[u8]>>>(&mut self, token: Option<D>) { ... }

}

Interface for CoAP messages common between requests, responses and other messages.

Required Methods

fn as_message(&self) -> &CoapMessage

Returns a reference to this message.

fn as_message_mut(&mut self) -> &mut CoapMessage

Returns a mutable reference to this message.

Provided Methods

fn add_option(&mut self, option: CoapOption)

Add the supplied CoAP option to this message.

fn clear_options(&mut self)

Clear the list of options that were added to this message using add_option().

fn options_iter(&self) -> Iter<'_, CoapOption>

Returns an iterator over the options contained in this message.

fn type_(&self) -> CoapMessageType

Returns the CoAP message type (confirmable, non-confirmable, acknowledgement, rst) of this message.

fn set_type_(&mut self, type_: CoapMessageType)

Sets the CoAP message type (confirmable, non-confirmable, acknowledgement, rst) of this message.

fn code(&self) -> CoapMessageCode

Returns the message code of this message. To determine whether the message is a request or response, use CoapMessageCode::try_from() and match for the enum variants.

fn set_code<C: Into<CoapMessageCode>>(&mut self, code: C)

Sets the message code of this message.

fn mid(&self) -> Option<CoapMessageId>

Returns the CoAP message ID for this message.

fn set_mid(&mut self, mid: Option<CoapMessageId>)

Sets the CoAP message ID for this message.

fn data(&self) -> Option<&[u8]>

Returns a reference to the data/body of this message.

fn set_data<D: Into<Box<[u8]>>>(&mut self, data: Option<D>)

Sets the data/body of this message.

fn token(&self) -> Option<&[u8]>

Returns the message token.

fn set_token<D: Into<Box<[u8]>>>(&mut self, token: Option<D>)

Sets the message token.

Note that CoapSessionCommon::send_request() will automatically set the token to a random value if you don’t.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.