libcoap_rs
Module protocol

Various types that are specified and defined in the CoAP standard and its extensions.

Enums

CoapContentFormat

Various content formats that can be used for CoAP requests.

CoapMatch

Representation of a CoAP match expression supplied in the If-Match option, see RFC 7252, Section 5.10.8.1.

CoapMessageCode

Representation of a CoAP message code. Can be a request code, a response code, or the empty message code.

CoapMessageType

CoAP message types as defined in RFC 7252, Section 3 and described in RFC 7252, Section 4.2 and 4.3.

CoapOptionType

CoAP option types as defined in RFC 7252, Section 5.10 and later CoAP extensions.

CoapRequestCode

Representation of a CoAP request/method code.

CoapResponseCode

Representation of a CoAP response code.

Type Aliases

Block

CoapOptionNum

CoapToken

ContentFormat

ETag

Echo

HopLimit

LocationPath

LocationQuery

MaxAge

NoResponse

Observe

Oscore

ProxyScheme

ProxyUri

RequestTag

Size

UriHost

UriPath

UriPort

UriQuery