libcoap_rs
Module session

Re-exports

pub use self::client::CoapClientSession;

pub use self::server::CoapServerSession;

Modules

client

server

Structs

CoapRequestHandle

A handle returned by CoAP sessions upon sending a request.

Enums

CoapSession

The representation of a CoAP session.

CoapSessionState

Representation of the states that a session can be in.

Traits

CoapSessionCommon

Trait for functions that are common between client and server sessions.