libcoap_rs::session

Enum CoapSession

pub enum CoapSession<'a> {
    Client(CoapClientSession<'a>),
    Server(CoapServerSession<'a>),
}

The representation of a CoAP session.

This enum only provides functionality that is common between clients and servers (by implementing CoapSessionCommon). If you require functionality specific to client- or server-side sessions, match this enum on the required session type.

Variants

Client(CoapClientSession<'a>)

Server(CoapServerSession<'a>)